Q3-0013_ElsDulis/requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
pydantic>=1.10.0
python-dotenv>=0.21.0

# Data Visualization
plotly>=5.14.0
seaborn>=0.12.0
matplotlib>=3.6.0

# Web Framework & UI
streamlit>=1.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Validation & Processing
polars>=0.19.0
pyarrow>=13.0.0

# AI & NLP
langchain>=0.0.300
openai>=1.3.0
anthropic>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
